backend/requirements.txt
# FastAPI and server dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0

# Data processing and scraping
requests>=2.26.0
beautifulsoup4>=4.9.3
ipython>=7.31.0

# OpenAI
openai>=1.0.0

# Optional but recommended
python-multipart>=0.0.5

pydantic==2.6.1
httpx==0.26.0
email-validator==2.1.0.post1
ujson==5.9.0 